+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cachestore/PatchFile.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2019, Oracle Corporation and/or its affiliates.  All rights reserved.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.impl;
+package com.oracle.weblogic.imagetool.cachestore;
 
 import java.io.IOException;
 import java.util.Collections;
