+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cli/cache/CacheOperation.java
@@ -9,7 +9,7 @@ import java.util.concurrent.Callable;
 
 import com.oracle.weblogic.imagetool.api.meta.CacheStore;
 import com.oracle.weblogic.imagetool.api.model.CommandResponse;
-import com.oracle.weblogic.imagetool.impl.meta.CacheStoreFactory;
+import com.oracle.weblogic.imagetool.cachestore.meta.CacheStoreFactory;
 import picocli.CommandLine.Model.CommandSpec;
 import picocli.CommandLine.Spec;
 import picocli.CommandLine.Unmatched;
