+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cli/menu/WLSInstallHelper.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Properties;
 
 import com.oracle.weblogic.imagetool.api.model.InstallerType;
-import com.oracle.weblogic.imagetool.impl.InstallerFile;
+import com.oracle.weblogic.imagetool.cachestore.InstallerFile;
 import com.oracle.weblogic.imagetool.logging.LoggingFacade;
 import com.oracle.weblogic.imagetool.logging.LoggingFactory;
 import com.oracle.weblogic.imagetool.util.Constants;
