+++ imagetool/src/test/java/com/oracle/weblogic/imagetool/cachestore/meta/FileCacheStoreTest.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2019, Oracle Corporation and/or its affiliates.  All rights reserved.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.impl.meta;
+package com.oracle.weblogic.imagetool.cachestore.meta;
 
 import com.oracle.weblogic.imagetool.api.meta.CacheStore;
 import org.junit.FixMethodOrder;
